+++ stepmania/src/RageSoundReader_PostBuffering.h
@@ -10,7 +10,7 @@ class RageSoundReader_PostBuffering: public RageSoundReader_Filter
 public:
 	RageSoundReader_PostBuffering( RageSoundReader *pSource );
 	RageSoundReader_PostBuffering *Copy() const { return new RageSoundReader_PostBuffering(*this); }
-	virtual int Read( int16_t *pBuf, int iFrames );
+	virtual int Read( float *pBuf, int iFrames );
 	virtual bool SetProperty( const RString &sProperty, float fValue );
 
 private:
