+++ stepmania/src/RageSoundReader_WAV.h
@@ -14,7 +14,7 @@ public:
 	void Close();
 	int GetLength() const;
 	int SetPosition( int iFrame );
-	int Read( int16_t *pBuf, int iFrames );
+	int Read( float *pBuf, int iFrames );
 	int GetSampleRate() const { return m_WavData.m_iSampleRate; }
 	unsigned GetNumChannels() const { return m_WavData.m_iChannels; }
 	int GetNextSourceFrame() const;
