+++ stepmania/src/RageSoundUtil.h
@@ -5,10 +5,10 @@
 
 namespace RageSoundUtil
 {
-	void Attenuate( int16_t *pBuf, int iSamples, float fVolume );
-	void Pan( int16_t *pBuffer, int iFrames, float fPos );
-	void Fade( int16_t *pBuffer, int iFrames, float fStartVolume, float fEndVolume  );
-	void ConvertMonoToStereoInPlace( int16_t *pBuffer, int iFrames );
+	void Attenuate( float *pBuf, int iSamples, float fVolume );
+	void Pan( float *pBuffer, int iFrames, float fPos );
+	void Fade( float *pBuffer, int iFrames, float fStartVolume, float fEndVolume  );
+	void ConvertMonoToStereoInPlace( float *pBuffer, int iFrames );
 };
 
 #endif
